Separated llg solver from main script.
